attestation: Separate the attestation manager from the measurement code
And have a clearer separation of duties.
This is a preparation for moving the DICE part to its separate crate.
